requirements_anaconda.txt
# Requirements ottimizzati per Anaconda
# Usa versioni compatibili con l'ambiente conda esistente

# Già installato in Anaconda - no need to reinstall
# Flask>=3.0.0
# matplotlib>=3.7.0  
# requests>=2.20.0

# Aggiungi solo pacchetti mancanti
openpyxl
schedule
requests-oauthlib

# Opzionali - installare solo se necessari
# pandas  # Già presente in Anaconda
# numpy   # Già presente in Anaconda
# Pillow  # Evitare se da problemi
